# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: EZNews API
  version: 1.0.0
  summary: 外部采集器写入（Ingest）+ Web/Android 读取查询 + TTS 语音合成
  description: |
    EZNews 是轻量、可自托管的新闻聚合与语音播报服务。

    ## 系统边界
    - **服务端不采集**：不爬虫、不解析 RSS、不抓网页。文章由**外部采集器**通过本文件的 `ingest` 接口写入。
    - **无推送**：无 WebSocket / SSE / 厂商推送。客户端主动 REST 拉取，配合 `cursor` 游标做增量同步。
    - **TTS 服务端合成**：客户端只播放音频，不做端侧朗读。

    ## 鉴权
    - 写接口（`POST /api/v1/ingest/*`）**必须**携带 `X-Api-Key` 请求头。
    - 读接口默认不鉴权，可由服务端配置 `auth.requireKeyForRead` 开启。

    ## 幂等
    `ingest` 接口幂等：以 `sourceId+externalId` 或规范化 `url` 为去重键；
    内容无变化时回执 `skipped` 且**不刷新 `updated_at`**（保证增量游标纯净）。
    因此对 `429 / 500 / 503` 可安全重试。

    ## 账号体系（v1.1）
    形态为「**游客可用 + 可选登录**」双态：**登录是增强，不是门槛**。

    - 业务接口（文章/源/音频）对游客**全部开放**，永不返回"需要登录"。
    - 仅 `/api/v1/me/**` 需要 Bearer Access Token。
    - Access Token = JWT(HS256)，2 小时，**服务端本地验签、高频路径零 DB 查询**。
    - Refresh Token = 32 字节不透明随机串，30 天滑动续期 + 轮换 + reuse 检测；库内仅存 SHA-256 哈希。
    - 改密/全局登出靠 `token_version` 递增，**无 Redis 黑名单**。
    - 登录后客户端自动调用 `POST /api/v1/me/merge` 幂等合并游客本地数据（无弹窗、无打断）。
    - **客户端 401 处理必须实现 single-flight（DEC-10，实现强制项）**：并发 401 只触发一次 refresh，其余请求挂起等待，否则 refresh token 会被并发轮换导致互踢。
    - **Web 官方仅支持同源部署（DEC-12）**：同域反代 + `httpOnly` Cookie。跨域部署不为官方支持（refreshToken 降级 localStorage，XSS 风险由部署者自担），不为其设计额外跨域认证方案。

    ## 时间格式
    所有时间字段以 ISO 8601 UTC（`YYYY-MM-DDTHH:mm:ss.sssZ`）传输；服务端内部以 epoch 毫秒存储。

  contact:
    name: EZNews Project
servers:
  - url: http://localhost:8080
    description: 本地自托管默认地址
  - url: https://{host}
...
security: []
paths:
...
  /api/v1/audio/tasks/{taskId}/retry:
    post:
      tags: [audio]
      operationId: retryAudioTask
      summary: 重试失败任务（重置 retry_count 并重新入队）
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '202':
          description: 已重新入队
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/AudioTaskDTO' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        采集器专用密钥。通过请求头传递：**不使用 query 参数**（避免出现在访问日志中）。
        服务端以常数时间比较校验；失败返回 401（不区分 key 不存在与 key 错误）。
        与账号体系**完全独立**：ingest 不使用 Bearer Token。

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        账号 Access Token（JWT HS256，2 小时）。
        - 业务接口为**可选**（`OptionalAuth`）：带则解析用户上下文，不带按游客放行。
        - `/api/v1/me/**` 为**必需**（`RequireAuth`）。
        - 401 时按 `code` 区分：`TOKEN_EXPIRED`（静默 refresh 后重放）与 `UNAUTHORIZED`（降级游客态）。
...
  schemas:
...
    Envelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            OK | BAD_REQUEST | VALIDATION_FAILED | UNAUTHORIZED | INVALID_API_KEY |
            INVALID_CREDENTIALS | TOKEN_EXPIRED | FORBIDDEN | NOT_FOUND | CONFLICT |
            PAYLOAD_TOO_LARGE | UNPROCESSABLE | RATE_LIMITED | TTS_UNAVAILABLE |
            TTS_QUEUE_FULL | SERVICE_BUSY | DB_BUSY | INTERNAL_ERROR

            其中 `TOKEN_EXPIRED` / `FORBIDDEN` 用于区分 401 与 403 的具体成因
            （见 securitySchemes.BearerAuth）；`TTS_UNAVAILABLE` / `TTS_QUEUE_FULL`
            为合成链路的两种临时态（503 / 429，均带 `Retry-After`），客户端应退避重试。
          example: OK
        message: { type: string, example: ok }
        requestId: { type: string, example: V1StGXR8_Z5jdHi6B-myT }
        data:
          description: 成功时为业务数据（对象或数组）；失败时可为 null 或含 details[]
...
    AudioTaskDTO:
      type: object
      required: [id, articleId, status]
      properties:
        id: { type: integer, format: int64 }
        articleId: { type: integer, format: int64 }
        voice: { type: string }
        speed: { type: number, format: float }
        status: { type: string, enum: [pending, processing, ready, failed] }
        audioId: { type: [integer, 'null'], format: int64 }
        provider: { type: [string, 'null'], enum: [tencent, ali, xunfei, 'null'] }
        retryCount: { type: integer }
        errorCode: { type: [string, 'null'] }
        errorMsg: { type: [string, 'null'] }
        textChars: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
...